+++ lib/workload/stateless/metadata_manager/src/library_manager/__init__.py
