+++ lib/workload/stateless/metadata_manager/src/library_manager/renderers.py
