+++ lib/workload/stateless/metadata_manager/src/library_manager/tests/case.py
@@ -1,0 +1,16 @@
+import logging
+
+from django.test import TestCase
+
+logger = logging.getLogger()
+logger.setLevel(logging.INFO)
+
+
+class LibraryUnitTestCase(TestCase):
+    def setUp(self) -> None:
+        logger.info("\n")
+        logger.info("-" * 64)
+        pass
+
+    def tearDown(self) -> None:
+        pass
