+++ lib/workload/stateless/metadata_manager/src/library_manager/wsgi.py
@@ -1,0 +1,5 @@
+import os
+
+from django.core.wsgi import get_wsgi_application
+
+application = get_wsgi_application()
