+++ lib/workload/stateless/metadata_manager/src/library_manager_proc/tests/case.py
@@ -1,0 +1,24 @@
+import logging
+
+from django.test import TestCase
+
+logger = logging.getLogger()
+logger.setLevel(logging.INFO)
+
+
+class LibraryProcUnitTestCase(TestCase):
+    def setUp(self) -> None:
+        logger.info("\n")
+        logger.info("-" * 64)
+        # some code construct that share across all test cases under lims package
+        # pass for now
+        pass
+
+    def tearDown(self) -> None:
+        # undo any construct done from setUp
+        # pass for now
+        pass
+
+
+class LibraryIntegrationTestCase(TestCase):
+    pass
