+++ lib/workload/stateless/metadata_manager/src/manage.py
@@ -6,7 +6,7 @@ import sys
 
 def main():
     """Run administrative tasks."""
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "metadata_manager.settings.local")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "library_manager.settings.local")
     try:
         from django.core.management import execute_from_command_line
     except ImportError as exc:
