--- lib/workload/stateless/metadata_manager/src/metadata_manager/tests/case.py
@@ -1,16 +1,0 @@
-import logging
-
-from django.test import TestCase
-
-logger = logging.getLogger()
-logger.setLevel(logging.INFO)
-
-
-class MetadataUnitTestCase(TestCase):
-    def setUp(self) -> None:
-        logger.info("\n")
-        logger.info("-" * 64)
-        pass
-
-    def tearDown(self) -> None:
-        pass
