--- lib/workload/stateless/metadata_manager/src/metadata_manager/wsgi.py
@@ -1,7 +1,0 @@
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "metadata_manager.settings.base")
-
-application = get_wsgi_application()
