+++ lib/workload/stateless/metadata_manager/src/migrate.py
@@ -7,8 +7,6 @@ import os
 
 from django.core.management import execute_from_command_line
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "metadata_manager.settings.base")
-
 
 def handler(event, context) -> str:
     execute_from_command_line(["./manage.py", "migrate"])
